TkClassWiz (#492)
* Split package

* Extensions

* DPI

* More extensions

* convert extensions

* Rename

* Split annotations

* gui_except

* finished splitting

* Pre and post extensions

* Reduce complexity
