+++ src/daf_gui/tkclasswiz/__init__.py
@@ -1,0 +1,6 @@
+"""
+TkClassWizard - Library for graphically defining objects based on class annotations.
+Works with Tkinter / TTKBootstrap.
+"""
+
+__version__ = "1.0.0"
